% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mecfa_final_functions.R
\name{bestECFA}
\alias{bestECFA}
\title{Pick the best-fit ECFA model(s)}
\usage{
bestECFA(object, max = TRUE)
}
\arguments{
\item{object}{A fit MECFA list.}

\item{max}{Was the objective function maximized (`TRUE`)?}
}
\description{
Pick the best-fit ECFA model(s)
}
