% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInitialModels.R
\name{createInitialModels}
\alias{createInitialModels}
\title{Create a series of initial models with relatively equal numbers of items per
factor.}
\usage{
createInitialModels(maxFactors, items)
}
\arguments{
\item{maxFactors}{Either a `character` fector of factor names or a `numeric` for the number of factors to consider.}

\item{items}{Either a `character` vector of item names or a `numeric` for number of items.}
}
\description{
Create a series of initial models with relatively equal numbers of items per
factor.
}
