% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecfaSA.R
\docType{class}
\name{ecfaSA-class}
\alias{ecfaSA-class}
\title{An S4 class for ecfaSA}
\value{
An S4 object of class `ecfaSA`.
}
\description{
An S4 class for ecfaSA
}
\section{Slots}{

\describe{
\item{\code{algorithm}}{Which metaheuristic algorithm was used.}

\item{\code{function_call}}{The original function call.}

\item{\code{all_fit}}{A vector of the model fit values.}

\item{\code{best_fit}}{The best observed model fit value.}

\item{\code{best_model}}{A `lavaan` object of the final solution.}

\item{\code{best_syntax}}{A `character` vector of the final solution model syntax.}

\item{\code{initial_model}}{A `character` vector of the initial model syntax.}

\item{\code{number_of_factors}}{A `numeric` of the number of factors.}

\item{\code{runtime}}{A `difftime` object of the total run time of the function.}
}}

