% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratorySA.R
\docType{class}
\name{exploratorySA-class}
\alias{exploratorySA-class}
\title{An S4 class for exploratorySA}
\description{
An S4 class for exploratorySA
}
\section{Slots}{

\describe{
\item{\code{totalTime}}{A `difftime` object of the total time it took to run the algorithm.}

\item{\code{functionCall}}{The  `match.call` arguments used to run the algorithm.}

\item{\code{modelResults}}{A `list` of model results of class `ecfaSA`.}
}}

