% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_SA.R
\name{exploratorySA}
\alias{exploratorySA}
\title{Exploratory CFA using Simulated Annealing}
\usage{
exploratorySA(
  initialModels,
  originalData,
  maxSteps = 5000,
  fitStatistic = "cfi",
  maximize = TRUE,
  Kirkpatrick = TRUE,
  randomNeighbor = TRUE,
  lavaan.model.specs = list(model.type = "cfa", auto.var = TRUE, estimator = "default",
    ordered = NULL, int.ov.free = TRUE, int.lv.free = FALSE, std.lv = TRUE,
    auto.fix.first = FALSE, auto.fix.single = TRUE, auto.cov.lv.x = TRUE, auto.th = TRUE,
    auto.delta = TRUE, auto.cov.y = TRUE),
  maxChanges = 5,
  restartCriteria = "consecutive",
  maximumConsecutive = 25,
  temperature = function(currentStep, maxSteps) {     currentTemp <- ((maxSteps -
    (currentStep))/maxSteps)^7 },
  bifactor = FALSE,
  ...
)
}
\arguments{
\item{initialModels}{The initial model as a `character` vector with lavaan model syntax. Each starting model should be its own element in the vector.}

\item{originalData}{The original `data.frame` with variable names.}

\item{maxSteps}{Numeric value for maximum number of steps for the simulated annealing algorithm.}

\item{fitStatistic}{String indicating the fit statistic to use. Must be an appropriate fit statistic reported by `lavaan::fitstatistic`. Default is `'cfi'`.}

\item{maximize}{Logical. Is the chosen `fitStatistic` to be maximized? Default is `TRUE`.}

\item{Kirkpatrick}{Default is `TRUE` to use the Kirkpatrick goal function. Also accepts custom-defined goal functions.}

\item{randomNeighbor}{Default is `TRUE` to use the default random neighbor function. Also accepts custom-defined functions to specify how random neighbor models are generated.}

\item{lavaan.model.specs}{A list of named values that specify the model specifications as found in `lavaan::lavaan`. Default values match the values found for `lavaan::cfa`.}

\item{maxChanges}{An `integer` value greater than 1 setting the maximum number of parameters to change within randomNeighbor. .}

\item{restartCriteria}{Either "consecutive" to restart after maxConsecutiveSelection times with the same model chosen in a row, or a user-defined function.}

\item{maximumConsecutive}{A positive `integer` value used with restartCriteria.}

\item{temperature}{Either an acceptable `character` value, or a user-defined temperature function. The acceptable values are "linear", "quadratic", or "logistic". Defaults to a custom function.}

\item{bifactor}{Logical. Indicates if the latent model is a bifactor model. If `TRUE`, assumes that the last latent variable in the provided model syntax is the bifactor (i.e., all of the retained items will be set to load on the last latent variable). Ignored if `maxItems==FALSE`.}

\item{...}{Further arguments to be passed to other functions. Not implemented for any of the included functions.}
}
\description{
Exploratory CFA using Simulated Annealing
}
