% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperatures.R
\name{linearTemperature}
\alias{linearTemperature}
\title{A series of internal temperature functions.
Directly accessing these functions is not advised unless you know what you are doing.}
\usage{
linearTemperature(currentStep, maxSteps)
}
\description{
A series of internal temperature functions.
Directly accessing these functions is not advised unless you know what you are doing.
}
\keyword{internal}
