% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mecfa_final_functions.R
\name{modelTableToString}
\alias{modelTableToString}
\title{Print the model in string format}
\usage{
modelTableToString(object)
}
\arguments{
\item{object}{A fit MECFA list.}
}
\description{
Print the model in string format
}
