% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratorySA.R
\name{plot,exploratorySA,ANY-method}
\alias{plot,exploratorySA,ANY-method}
\title{Plot method for class `exploratorySA`}
\usage{
\S4method{plot}{exploratorySA,ANY}(x, y, ...)
}
\arguments{
\item{x, y}{An S4 object of class `exploratorySA`.}

\item{...}{Not used.}
}
\description{
Plot method for class `exploratorySA`
}
