% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecfaSA.R
\name{show,ecfaSA-method}
\alias{show,ecfaSA-method}
\title{Print method for class `ecfaSA`}
\usage{
\S4method{show}{ecfaSA}(object)
}
\arguments{
\item{object}{An S4 object of class `ecfaSA`.}
}
\description{
Print method for class `ecfaSA`
}
