% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratorySA.R
\name{show,exploratorySA-method}
\alias{show,exploratorySA-method}
\title{Print method for class `exploratorySA`}
\usage{
\S4method{show}{exploratorySA}(object)
}
\arguments{
\item{object}{An S4 object of class `exploratorySA`.}
}
\description{
Print method for class `exploratorySA`
}
