% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratorySA.R
\name{summary,exploratorySA-method}
\alias{summary,exploratorySA-method}
\title{Summary method for S4 object `exploratorySA`}
\usage{
\S4method{summary}{exploratorySA}(object)
}
\arguments{
\item{object}{An S4 object of class `exploratorySA`

export}
}
\description{
Summary method for S4 object `exploratorySA`
}
